+++ deps/pulled.json
@@ -2,8 +2,8 @@
   "canisters": {
     "rdmx6-jaaaa-aaaaa-aaadq-cai": {
       "name": "internet_identity",
-      "wasm_hash": "4fae35068ed79887f8de96823fc38cd0448173d1c4d2c5e98ad39f3fdec1b715",
-      "wasm_hash_download": "4fae35068ed79887f8de96823fc38cd0448173d1c4d2c5e98ad39f3fdec1b715",
+      "wasm_hash": "b8716ccde2e13e785baa2f983cf1657a00a7459394c6f08ca398b7b6ccfe391b",
+      "wasm_hash_download": "b8716ccde2e13e785baa2f983cf1657a00a7459394c6f08ca398b7b6ccfe391b",
       "init_guide": "Use '(null)' for sensible defaults. See the candid interface for more details.",
       "init_arg": null,
       "candid_args": "(opt InternetIdentityInit)",
