+++ package.json
@@ -27,8 +27,10 @@
     "@headlessui/react": "^1.7.18",
     "@heroicons/react": "^2.1.1",
     "@hookform/resolvers": "^3.3.4",
+    "@material-tailwind/react": "^2.1.9",
     "@mui/icons-material": "^5.15.13",
     "@mui/material": "^5.15.13",
+    "@ramonak/react-progress-bar": "^5.2.0",
     "@react-hook/media-query": "^1.1.1",
     "@reduxjs/toolkit": "^2.2.1",
     "@vidstack/react": "^1.9.7",
