+++ src/frontend_canister/assets/main.css
@@ -17,6 +17,16 @@
   src: url("../assets/Fonts/QuickSand.ttf");
 }
 
+body {
+  font-family: "poppins";
+}
+
+
+.lightfont{
+  font-family: "nunitoSans" !important;
+}
+
+
 /* Hide the up and down arrows */
 input[type="number"]::-webkit-inner-spin-button,
 input[type="number"]::-webkit-outer-spin-button {
@@ -32,12 +42,33 @@ input[type="number"] {
 }
 
 .sidebar_icon_size {
-  @apply w-6 h-6 object-contain;
+  @apply object-contain w-6 h-6;
 }
 
 .sidebar_text_style {
-  @apply ml-4 font-serif font-medium text-xs xl:text-base;
+  @apply ml-4 font-serif text-xs font-medium xl:text-base;
 }
 .input_foucs_border {
   @apply border border-[#ded9fa] focus:outline-none p-2 focus:border-[#9189c4];
+}
+
+.dashboard_cap_gradient {
+  background: rgb(112, 85, 244);
+  background: linear-gradient(90deg, rgb(112, 85, 244) 0%, rgb(199, 189, 243) 100%);
+}
+
+.dashboard_cap_gradient2 {
+  background: rgb(101,154,251);
+background: linear-gradient(90deg, rgba(101,154,251,1) 0%, rgba(211,229,255,1) 100%);
+}
+
+
+.react-calendar{
+  border: none !important;
+  border-radius: 15px !important;
+  padding: 1rem !important;  
+}
+
+abbr:where([title]){
+  text-decoration: none !important;
 }
